[ft 166057229] add protection to comment

+++ server/controllers/comments.js
@@ -93,6 +93,23 @@ class Comments {
         return false;
       }
       const { id } = req.params;
+      const commentOwner = await Comment.findOne({
+        where: {
+          id,
+        }
+      });
+      if (commentOwner.userId !== req.user.id) {
+        return res.status(401).json({
+          status: 401,
+          message: 'This is not your comment'
+        });
+      }
+      if (!commentOwner) {
+        return res.status(401).json({
+          status: 404,
+          message: 'Comment not found'
+        });
+      }
       const result = await Comment.destroy({ where: { id, }, returning: true });
 
       if (result > 0) {
